refactor(sdk/analyze): split patterns.rs into per-family detector submodules
patterns.rs had grown to 1564 lines bundling ~10 independent detectors plus
shared plumbing. Split the detector families into cohesive submodules under
patterns/, leaving the shared plumbing and the detect_patterns orchestrator
in the parent:

  patterns.rs        755  consts, tool-name helpers, DetectPatternsOptions,
                          detect_patterns, ref structs, ContentIndex,
                          flatten/dedup/cost helpers, detect_streaks/StreakOp,
                          graph event helpers, build_summaries
  patterns/streaks.rs    372  retry / failure / cancellation (graph + flat) + signatures
  patterns/edits.rs      157  edit-revert, edit-heavy
  patterns/compaction.rs 150  compaction-loss + window summary
  patterns/skills.rs     157  skill recall-dup, pruning-protection, system-prompt-tax

Pure code movement. Each submodule reaches the parent's shared plumbing via
`use super::*`. The only non-movement edits are widening GraphStatusPatterns /
detect_graph_status_patterns_for_session / detect_compaction_losses from
private to pub(super) so the parent orchestrator can call across the boundary.
detect_patterns's body is unchanged. All 992 tests pass, clippy clean, build
warning-free; the moved function bodies are byte-identical to the originals.
